♻️ refactor(sync): migrate shelf_books to descriptor catalog
